feat: rename header to tab-section, using styles from theme
