+++ src/layouts/tab-section/index.ts
@@ -1,0 +1,1 @@
+export * from './tab-section';
